Removed custom wallpaper. Code cleanup.

+++ zero/src/main/java/com/zero/zerolivewallpaper/wallpaper/MyRenderer.java
@@ -342,7 +342,7 @@ class MyRenderer implements GLSurfaceView.Renderer {
     }
 
     void setOffset(float offset) {
-        this.offset = (double) offset;
+        this.offset = offset;
     }
 
     private void deployFallbackWallpaper() {
